pyproject.toml
[tool.poetry]
name = "healthatom_test"
version = "v0.0.1"
description = ""
authors = ["Juan Pablo Cano"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
loguru = "^0.7.2"
sqlalchemy = "^2.0.25"
pydantic = "^2.5.3"
click = "^8.1.7"

[tool.poetry.group.dev.dependencies]
black = ">=22.12.0,<22.13.0"
flake8 = ">=6.0.0,<6.1.0"
mypy = ">=1.7.1,<1.8.0"
isort = ">=5.11.4,<5.12.0"
pydocstyle = ">=6.2.3,<6.3.0"
python-dotenv = "^1.0.0"
pytest = ">=7.2.1,<7.3.0"
pytest-mock = ">=3.10.0,<3.11.0"
pytest-cov = ">=4.0.0,<4.1.0"
types-requests = "^2.31.0.20240106"
responses = ">=0.23.1,<0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length=100
target-version=['py310']
skip-string-normalization=true
exclude='''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | env | venv | .venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''
